refactor(authentication): authz teams routes initialization
